[Gold III] Title: Teamwork, Time: 328 ms, Memory: 55456 KB -BaekjoonHub

+++ 백준/Gold/16765. Teamwork/Teamwork.java
@@ -2,7 +2,6 @@
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.Arrays;
 import java.util.StringTokenizer;
 
 public class Main {
@@ -14,18 +13,18 @@ public class Main {
         int N = Integer.parseInt(st.nextToken());
         int K = Integer.parseInt(st.nextToken());
 
-        long[] cows = new long[N + 1];
+        int[] cows = new int[N + 1];
         for (int i = 1; i <= N; i++) {
             cows[i] = Integer.parseInt(br.readLine());
         }
 
-        long[][] dp = new long[K + 1][N + 1];
+        int[][] dp = new int[K + 1][N + 1];
 
-        long max = 0;
+        int max = 0;
         for (int i = 1; i <= N; i++) {
 
-            long nowMax = 0;
-            long maxS = cows[i];
+            int nowMax = 0;
+            int maxS = cows[i];
 
             for (int k = 1; k <= K; k++) {
                 if (i - k < 0) {
@@ -34,10 +33,9 @@ public class Main {
                 int beforeIndex = i - k;
 
                 maxS = Math.max(cows[beforeIndex+1], maxS);
-                long totalS = maxS * k;
+                int totalS = maxS * k;
 
-//                System.out.println("totalS = " + totalS);
-                long from = dp[0][beforeIndex];
+                int from = dp[0][beforeIndex];
                 dp[k][i] = Math.max(dp[k][i], from + totalS);
                 nowMax = Math.max(nowMax, dp[k][i]);
             }
@@ -45,13 +43,7 @@ public class Main {
             max = Math.max(max, nowMax);
 
         }
-//        for(int i = 1;i<=K;++i) {
-//            System.out.println(Arrays.toString(dp[i]));
-//        }
-
         System.out.println(max);
-
-
     }
 
 }
